Compaq/Alpha debugging on waterbug: silencing gcc warnings, and one bugfix in CMRebalance(), begin[] and end[] distributions weren't being copied.
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/trunk/infernal@830 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/alphabet.c
@@ -156,7 +156,7 @@ DigitizeSequence(char *seq, int L)
   for (i = 0; i < L; i++) {
     if (isgap(seq[i])) dsq[i+1] = DIGITAL_GAP; 
     else {
-      c = toupper(seq[i]);
+      c = toupper((int) seq[i]);
       if (c == 'T') c = 'U';	/* it's RNA, dammit. */
       dsq[i+1] = SymbolIndex(c);
     }
